feat: replace leaflet/article permission hooks with unified useGenAiUserPermission
- Create useGenAiUserPermission hook checking the genai_user role
- Remove useArticleGeneratorPermission from useArticles.ts
- Remove useLeafletUploadPermission from useLeaflet.ts
- Migrate ArticleGenerationForm, LeafletGeneratorPage, MarketingFeedbackPage
  to use the new unified hook
- Consolidate two separate permission checks in MarketingFeedbackPage to one
- Update Sidebar to check genai_user instead of leaflet_manager/article_generator
- Add genai_user role to mock auth seed user
