+++ frontend/src/api/hooks/useGenAiUserPermission.ts
@@ -1,0 +1,18 @@
+import { useMsal } from '@azure/msal-react';
+import { shouldUseMockAuth } from '../../config/runtimeConfig';
+import { mockAuthService } from '../../auth/mockAuth';
+
+export const useGenAiUserPermission = (): boolean => {
+  const { accounts } = useMsal();
+
+  if (shouldUseMockAuth()) {
+    const user = mockAuthService.getUser();
+    return !!(Array.isArray(user?.roles) && user?.roles.includes('genai_user'));
+  }
+
+  const account = accounts[0];
+  if (!account) return false;
+  const claims = account.idTokenClaims as Record<string, unknown> | undefined;
+  const roles = claims?.['roles'];
+  return Array.isArray(roles) && roles.includes('genai_user');
+};
