+++ frontend/src/features/articles/ArticleGenerationForm.tsx
@@ -1,7 +1,8 @@
 import { useState } from 'react';
 import { Loader2 } from 'lucide-react';
 import { GenerateArticleRequest } from '../../api/generated/api-client';
-import { useGenerateArticleMutation, useArticleGeneratorPermission } from '../../api/hooks/useArticles';
+import { useGenerateArticleMutation } from '../../api/hooks/useArticles';
+import { useGenAiUserPermission } from '../../api/hooks/useGenAiUserPermission';
 
 interface ArticleGenerationFormProps {
   onArticleCreated: (articleId: string) => void;
@@ -21,7 +22,7 @@ const LENGTH_OPTIONS = [
 ];
 
 export default function ArticleGenerationForm({ onArticleCreated }: ArticleGenerationFormProps) {
-  const canGenerate = useArticleGeneratorPermission();
+  const canGenerate = useGenAiUserPermission();
   const { mutate: generate, isPending, error } = useGenerateArticleMutation();
 
   const [topic, setTopic] = useState('');
